Move select_method to MethodProperties
