+++ src/inmanta/protocol/__init__.py
@@ -55,9 +55,9 @@ to REST transport with Tornado, together with the code in :module:`~inmanta.serv
 
 from inmanta.protocol import methods, methods_v2
 from inmanta.protocol.auth.auth import decode_token, encode_token
-from inmanta.protocol.common import Response, Result, gzipped_json, json_encode
+from inmanta.protocol.common import Response, Result, VersionMatch, gzipped_json, json_encode
 from inmanta.protocol.decorators import handle, method, typedmethod
-from inmanta.protocol.endpoints import Client, SessionClient, SessionEndpoint, SyncClient, TypedClient, VersionMatch
+from inmanta.protocol.endpoints import Client, SessionClient, SessionEndpoint, SyncClient, TypedClient
 
 __all__ = [
     "Response",
